Apply default nodeSelector in helm chart
We only build the operator for amd64/linux currently, so the operator
pod should select for nodes of that type.

Since we currently support 1.13, and the selector labels changed (left
beta) between 1.13 & 1.14, there is template ugliness required to ensure
the selectors can be customized (and removed if desired) while also
getting the correct beta vs. not label by default.

Signed-off-by: John Strunk <jstrunk@redhat.com>